Rename post directory to log
