+++ lib/public/components/Log/logCollapsed.js
