+++ lib/public/components/Log/logExpanded.js
