+++ lib/public/views/Logs/Details/logTreeViewComponent.js
@@ -14,8 +14,8 @@
 import { h } from '/js/src/index.js';
 import errorAlert from '../../../components/common/errorAlert.js';
 import spinner from '../../../components/common/spinner.js';
-import { logCollapsed } from '../../../components/Post/logCollapsed.js';
-import { logExpanded } from '../../../components/Post/logExpanded.js';
+import { logCollapsed } from '../../../components/Log/logCollapsed.js';
+import { logExpanded } from '../../../components/Log/logExpanded.js';
 import scrollTo from '../../../utilities/scrollTo.js';
 import { frontLink } from '../../../components/common/navigation/frontLink.js';
